Tried to apply lemouth's corrections request:
if the Les Houches info is not provided, the parameters have to be included in a freshly created
block. This is implemented and will work fine for scalar parameters where the block FRBlock is assigned. By the way this should become
PyRulesBlock (FRBlock stands for FeynRulesBlock :p). However, for tensor parameters, they must lie in their own block -> a freshly unique name
must be generated. the allowsummation option must disappear. I don’t want it because it is ugly ;) Tensorparameters don’t have any orderblock
attribute, as the orderblocks are ntuples with the line number, column number, etc...